Added sub navigation and styles
Constructed a basic sub navigation that helps the user switch between
views for issues list and reports. Navigation items are grouped under
the name of a scheme.

We have to test the navigation as it's developed based on the assumption
that the app presents data (issues list) per Scheme (requires the user
to choose a scheme first).
I think this may not be true and the user might want to view information
and access data based on different needs / tasks / priorities that are
'independent' of scheme or structure of 'entity model'.